requirements-dev.txt
pytest==8.1.1
pytest-cov==5.0.0
moto[dynamodb]==5.0.5
ruff==0.4.1
mypy==1.9.0
boto3-stubs[dynamodb]==1.34.69
